Search [has auth and some search capabilities]

+++ server/routes/orders.js
@@ -5,7 +5,12 @@ const orderRoute = express.Router();
 
 orderRoute.get('/', async(req, res, next) => {
     try {
-        res.send(await Order.findAll())
+        const admin = req.user && req.user.class === 'admin';
+        if (admin) {
+            res.send(await Order.findAll());
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
@@ -14,13 +19,19 @@ orderRoute.get('/', async(req, res, next) => {
 
 orderRoute.get('/:id', async(req, res, next) => {
     try {
-        res.send(await Order.findByPk(req.params.id, {
-            include: [{
-                model: Item
-            }, {
-                model: User
-            }]
-        }));
+        const admin = req.user && req.user.class === 'admin';
+        const ownUser = req.user.id === req.params.id * 1;
+        if (admin || ownUser) {
+            res.send(await Order.findByPk(req.params.id, {
+                include: [{
+                    model: Item
+                }, {
+                    model: User
+                }]
+            }));
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
@@ -29,8 +40,14 @@ orderRoute.get('/:id', async(req, res, next) => {
 
 orderRoute.post('/', async(req,res,next) => {
     try {
-        const order = await Order.create(req.body);
-        res.status(201).send(order);
+        const admin = req.user && req.user.class === 'admin';
+        const notUser = !req.user || req.user.class === 'guest';
+        if(admin || notUser) {
+            const order = await Order.create(req.body);
+            res.status(201).send(order);
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
@@ -39,8 +56,14 @@ orderRoute.post('/', async(req,res,next) => {
 
 orderRoute.delete('/:id', async(req, res, next) => {
     try {
-        await Order.destroy({ where: { id: req.params.id } });
-        res.sendStatus(204);
+        const admin = req.user && req.user.class === 'admin';
+        const ownUser = req.user.id === req.params.id * 1;
+        if (admin || ownUser) {
+            await Order.destroy({ where: { id: req.params.id } });
+            res.sendStatus(204);
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
@@ -49,9 +72,15 @@ orderRoute.delete('/:id', async(req, res, next) => {
 
 orderRoute.put('/:id', async(req, res, next) => {
     try {
-        const order = await Order.findByPk(req.params.id);
-        await Order.update(req.body);
-        res.send(order);
+        const admin = req.user && req.user.class === 'admin';
+        const ownUser = req.user.id === req.params.id * 1;
+        if (admin || ownUser) {
+            const order = await Order.findByPk(req.params.id);
+            await Order.update(req.body);
+            res.send(order);
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
