+++ server/routes/users.js
@@ -4,8 +4,13 @@ const userRoute = express.Router();
 
 userRoute.get('/', async(req, res, next) => {
     try {
-        if (req.user && req.user.class === 'admin') {
-            res.send(await User.findAll())
+        const admin = req.user && req.user.class === 'admin';
+        if (admin) {
+            res.send(await User.findAll({
+                attributes: {
+                    exclude: ['password', 'id']
+                }
+            }))
         } else {
             res.sendStatus(403);
         }
@@ -17,17 +22,21 @@ userRoute.get('/', async(req, res, next) => {
 
 userRoute.get('/:id', async(req, res, next) => {
     try {
-        if (req.user && req.user.class === 'admin' || req.user && req.user.id === req.params.id) {
+        const admin = req.user && req.user.class === 'admin';
+        const ownUser = req.user.id === req.params.id * 1;
+        if (admin || ownUser) {
             res.send(await User.findByPk(req.params.id, {
                 include: [{
                     model: Order,
                     include: [{ model: Item }]
-                }]
+                }],
+                attributes: {
+                    exclude: ['password', 'id']
+                }
             }));
         } else {
             res.sendStatus(403);
         }
-
     }
     catch(err) {
         console.log(err);
@@ -36,8 +45,14 @@ userRoute.get('/:id', async(req, res, next) => {
 
 userRoute.post('/', async(req,res,next) => {
     try {
-        const user = await User.create(req.body);
-        res.status(201).send(user);
+        const admin = req.user && req.user.class === 'admin';
+        const notUser = !req.user || req.user.class === 'guest';
+        if(admin || notUser) {
+            const user = await User.create(req.body);
+            res.status(201).send(user);
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
@@ -47,8 +62,14 @@ userRoute.post('/', async(req,res,next) => {
 
 userRoute.delete('/:id', async(req, res, next) => {
     try {
-        await User.destroy({ where: { id: req.params.id } });
-        res.sendStatus(204);
+        const admin = req.user && req.user.class === 'admin';
+        const ownUser = req.user.id === req.params.id * 1;
+        if (admin || ownUser) {
+            await User.destroy({ where: { id: req.params.id } });
+            res.sendStatus(204);
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
@@ -57,9 +78,15 @@ userRoute.delete('/:id', async(req, res, next) => {
 
 userRoute.put('/:id', async(req, res, next) => {
     try {
-        const user = await User.findByPk(req.params.id);
-        await User.update(req.body);
-        res.send(user);
+        const admin = req.user && req.user.class === 'admin';
+        const ownUser = req.user.id === req.params.id * 1;
+        if (admin || ownUser) {
+            const user = await User.findByPk(req.params.id);
+            await User.update(req.body);
+            res.send(user);
+        } else {
+            res.sendStatus(403);
+        }
     }
     catch(err) {
         console.log(err);
